Fix RideMapper
